+++ pkg/validator/alert.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/fusakla/promruval/v2/pkg/prometheus"
+	"github.com/fusakla/promruval/v2/pkg/unmarshaler"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/model/rulefmt"
 	"gopkg.in/yaml.v3"
@@ -30,7 +31,7 @@ func (h forIsNotLongerThan) String() string {
 	return fmt.Sprintf("alert `for` is not longer than `%s`", h.limit)
 }
 
-func (h forIsNotLongerThan) Validate(rule rulefmt.Rule, _ *prometheus.Client) []error {
+func (h forIsNotLongerThan) Validate(_ unmarshaler.RuleGroup, rule rulefmt.Rule, _ *prometheus.Client) []error {
 	if rule.For != 0 && rule.For > h.limit {
 		return []error{fmt.Errorf("alert has `for: %s` which is longer than the specified limit of %s", rule.For, h.limit)}
 	}