+++ pkg/validator/expression_helpers.go
@@ -2,8 +2,10 @@ package validator
 
 import (
 	"fmt"
+
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/promql/parser"
+	"golang.org/x/exp/slices"
 )
 
 func getExpressionUsedLabels(expr string) ([]string, error) {
@@ -28,53 +30,63 @@ func getExpressionUsedLabels(expr string) ([]string, error) {
 		}
 		return nil
 	})
-	return usedLabels, nil
+	slices.Sort(usedLabels)
+	return slices.Compact(usedLabels), nil
 }
 
-func getExpressionSelectors(expr string) ([]string, error) {
+func getExpressionVectorSelectors(expr string) ([]*parser.VectorSelector, error) {
 	promQl, err := parser.ParseExpr(expr)
 	if err != nil {
-		return []string{}, fmt.Errorf("failed to parse expression `%s`: %s", expr, err)
+		return []*parser.VectorSelector{}, fmt.Errorf("failed to parse expression `%s`: %s", expr, err)
 	}
-	var selectors []string
+	var selectors []*parser.VectorSelector
 	parser.Inspect(promQl, func(n parser.Node, ns []parser.Node) error {
 		switch v := n.(type) {
 		case *parser.VectorSelector:
-			s := &parser.VectorSelector{Name: v.Name, LabelMatchers: v.LabelMatchers}
-			selectors = append(selectors, s.String())
+			selectors = append(selectors, &parser.VectorSelector{Name: v.Name, LabelMatchers: v.LabelMatchers})
 		}
 		return nil
 	})
 	return selectors, nil
 }
 
-type VectorSelectorWithMetricName struct {
-	Vector     *parser.VectorSelector
-	MetricName string
+func getVectorSelectorMetricName(selector *parser.VectorSelector) string {
+	if selector.Name == "" {
+		for _, m := range selector.LabelMatchers {
+			if m.Name == "__name__" && m.Type == labels.MatchEqual {
+				return m.Value
+			}
+		}
+	}
+	return selector.Name
 }
 
-func getExpressionMetricsNames(expr string) ([]VectorSelectorWithMetricName, error) {
-	promQl, err := parser.ParseExpr(expr)
+// MetricWithVectorSelector is a struct that contains a metric name and a vector selector where it is used, to give a context, in the error messages.
+type MetricWithVectorSelector struct {
+	VectorSelector     *parser.VectorSelector
+	Name string
+}
+
+func getExpressionMetrics(expr string) ([]MetricWithVectorSelector, error) {
+	metrics := []MetricWithVectorSelector{}
+	vectorSelectors, err := getExpressionVectorSelectors(expr)
 	if err != nil {
-		return []VectorSelectorWithMetricName{}, fmt.Errorf("failed to parse expression `%s`: %s", expr, err)
+		return metrics, err
 	}
-	var vectors []VectorSelectorWithMetricName
-	parser.Inspect(promQl, func(n parser.Node, ns []parser.Node) error {
-		switch v := n.(type) {
-		case *parser.VectorSelector:
-			metricName := getMetricNameFromLabels(v.LabelMatchers)
-			vectors = append(vectors, VectorSelectorWithMetricName{Vector: v, MetricName: metricName})
-		}
-		return nil
-	})
-	return vectors, nil
+	for _, s := range vectorSelectors {
+		metrics = append(metrics, MetricWithVectorSelector{VectorSelector: s, Name: getVectorSelectorMetricName(s)})
+	}
+	return metrics, nil
 }
 
-func getMetricNameFromLabels(labels []*labels.Matcher) string {
-	for _, l := range labels {
-		if l.Name == "__name__" {
-			return l.Value
-		}
+func getExpressionSelectors(expr string) ([]string, error) {
+	vectorSelectors, err := getExpressionVectorSelectors(expr)
+	if err != nil {
+		return []string{}, err
 	}
-	return ""
+	var selectors []string
+	for _, s := range vectorSelectors {
+		selectors = append(selectors, s.String())
+	}
+	return selectors, nil
 }